made runLimitslxplus more robust

+++ runLimitslxplus.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 
+REMOTEDIR=/afs/cern.ch/user/j/jhugon/work/private/stats/CMSSW_5_2_5/stats/
+if ! `ssh lxplus touch $REMOTEDIR/touchfile.txt`; then
+  echo "Error: Remote directory $REMOTEDIR doesn't exist or isn't writable"
+  exit
+fi
+
 rm -f shapes/*
 rm -f statsCards/*
 rm -f statsInput/*
 rm -f statsOutput/*
-
-REMOTEDIR=/afs/cern.ch/user/j/jhugon/work/private/stats/CMSSW_5_2_5/stats/
 
 nice ./makeCards.py
 echo "Removing files in lxplus:$REMOTEDIR"
@@ -13,7 +17,7 @@ ssh lxplus "cd /tmp/jhugon/; rm -rf $REMOTEDIR/*;echo \"Contents of dir: \`ls $R
 echo "Copying input files to lxplus..."
 scp statsCards/* lxplus:/afs/cern.ch/user/j/jhugon/work/private/stats/CMSSW_5_2_5/stats/.
 echo "Running combine on lxplus..."
-ssh lxplus "cd $REMOTEDIR; eval \`scramv1 runtime -sh\`; nice bash notlxbatch.sh;"
+ssh lxplus "cd $REMOTEDIR; eval \`scramv1 runtime -sh\`;nice bash notlxbatch.sh;"
 echo "Copying output files from lxplus..."
 scp lxplus:$REMOTEDIR/*.out statsInput/.
